Remove NATS dependency entirely. Refactor logic to export to workspace and import from airtable. Additionally, make the discovery of a view in the form  more robust
