+++ src/app/api/v1/data_exports/controllers.rs
@@ -1,7 +1,5 @@
 //! Controllers for the data exports API.
 
-use std::sync::Arc;
-
 use axum::extract::{Path, State};
 use axum::http::StatusCode;
 use axum::response::Response;
@@ -10,14 +8,13 @@ use chrono::Utc;
 use tokio::task;
 use uuid::Uuid;
 
-use crate::app::api::v1::data_exports::export_users_to_workspace::{
-    export_task, ExportUsersToWorkspaceTaskParams,
-};
+use super::workspace::policies::{EmailPolicy, PasswordPolicy};
+use super::workspace::{export_task, ExportParams};
+use super::ExportServices;
 use crate::app::api::v1::data_exports::requests::ExportUsersToWorkspaceRequest;
 use crate::app::api::v1::data_exports::responses::ExportUsersToWorkspaceResponse;
 use crate::app::api_response;
 use crate::app::errors::AppError;
-use crate::app::state::Services;
 use crate::services::auth::AuthData;
 use crate::services::storage::jobs::CreateJobBuilder;
 use crate::services::storage::types::{ExportDesination, JobData, JobDetails, JobType};
@@ -47,16 +44,11 @@ use crate::services::storage::ExecOptsBuilder;
     ),
 )]
 pub async fn export_users_to_workspace(
-    State(ctx): State<Arc<Services>>,
+    State(services): State<ExportServices>,
     Path(project_cycle_id): Path<Uuid>,
     Extension(auth): Extension<AuthData>,
     Json(request): Json<ExportUsersToWorkspaceRequest>,
 ) -> Result<Response, AppError> {
-    let storage_layer = ctx.storage_layer.clone();
-    let workspace = ctx.workspace.clone();
-    let nats = ctx.nats.clone();
-    let mail = ctx.mail.clone();
-
     let current_time = Utc::now();
     let time_only = current_time.format("%H:%M:%S").to_string();
 
@@ -72,28 +64,55 @@ pub async fn export_users_to_workspace(
         })
         .build()?;
 
-    let job_id = storage_layer
+    let job_id = services
+        .storage_layer
         .create_job(Some(project_cycle_id), data, &mut ExecOptsBuilder::default().build()?)
         .await?;
 
-    let subject = format!("pantheon.export.cancel.{}", job_id);
-    log::info!("{job_id}");
+    let email_policy = EmailPolicy::from(&request);
+    let password_policy = PasswordPolicy::from(&request);
 
-    let mut subscriber = nats.subscribe(subject).await?;
-    subscriber.unsubscribe_after(1).await?;
+    let already_exported = services
+        .storage_layer
+        .fetch_exported_volunteer_details_by_project_cycle(
+            project_cycle_id,
+            &mut ExecOptsBuilder::default().build()?,
+        )
+        .await?
+        .iter()
+        .map(|v| v.volunteer_id)
+        .collect::<Vec<Uuid>>();
 
-    let principal = auth.email()?;
+    let volunteers = if request.skip_users_on_conflict {
+        log::info!("Skipping users that have already been exported");
+        request
+            .volunteers
+            .into_iter()
+            .filter(|v| !already_exported.contains(&v.volunteer_id))
+            .collect()
+    } else {
+        for v in &request.volunteers {
+            if already_exported.contains(&v.volunteer_id) {
+                log::error!("One or more users have already been exported");
+                return Ok(api_response::error(
+                    StatusCode::BAD_REQUEST,
+                    "One or more users have already been exported",
+                ));
+            }
+        }
+        request.volunteers
+    };
 
-    let params = ExportUsersToWorkspaceTaskParams {
-        subscriber,
-        request,
-        principal,
+    let params = ExportParams {
         job_id,
-        project_cycle_id,
+        email_policy,
+        password_policy,
+        principal: auth.email()?,
+        volunteers,
     };
 
     task::spawn(async move {
-        let _ = export_task(storage_layer.clone(), workspace.clone(), mail, params).await;
+        let _ = export_task(&services, params).await;
     });
 
     Ok(api_response::success(StatusCode::OK, ExportUsersToWorkspaceResponse { job_id })?)
