+++ src/app/api/v1/data_exports/mod.rs
@@ -1,18 +1,35 @@
 //! Data Exports API.
 
 mod controllers;
-mod export_users_to_workspace;
 mod requests;
 mod responses;
+mod workspace;
 
 use std::sync::Arc;
 
+use axum::extract::FromRef;
 use axum::middleware::from_fn_with_state;
 use axum::{routing, Router};
 use utoipa::OpenApi;
 
 use crate::app::api::middleware::make_rbac;
 use crate::app::state::Services;
+
+struct ExportServices {
+    pub storage_layer: Arc<dyn crate::services::storage::StorageService>,
+    pub workspace: Arc<dyn crate::services::workspace::WorkspaceService>,
+    pub mail: Arc<dyn crate::services::mail::MailService>,
+}
+
+impl FromRef<Arc<Services>> for ExportServices {
+    fn from_ref(ctx: &Arc<Services>) -> Self {
+        Self {
+            storage_layer: ctx.storage_layer.clone(),
+            workspace: ctx.workspace.clone(),
+            mail: ctx.mail.clone(),
+        }
+    }
+}
 
 /// Documents the API for data exports
 #[derive(OpenApi)]
@@ -34,6 +51,7 @@ pub async fn build(ctx: Arc<Services>) -> Router<()> {
 
     Router::new()
         .route("/:project_cycle_id/workspace", export_users_to_workspace)
+        // .route("/x", routing::post(workspace::c))
         .route_layer(from_fn_with_state(ctx.clone(), guard1))
         .with_state(ctx.clone())
 }
