+++ src/app/api/v1/data_imports/controllers.rs
@@ -8,7 +8,8 @@ use axum::Json;
 use chrono::Utc;
 use tokio::task;
 
-use super::{import_airtable_base_task, ImportAirtableBaseTaskParams};
+use super::ImportServices;
+use crate::app::api::v1::data_imports::airtable::{import_task, ImportParams};
 use crate::app::api::v1::data_imports::requests::ImportAirtableBase;
 use crate::app::api::v1::data_imports::responses::AvailableBases;
 use crate::app::api_response;
@@ -53,13 +54,12 @@ pub async fn list_available_airtable_bases(
     )
 ]
 pub async fn import_airtable_base(
-    State(ctx): State<Arc<Services>>,
+    State(services): State<ImportServices>,
     Path(base_id): Path<String>,
     Json(payload): Json<ImportAirtableBase>,
 ) -> Result<Response, AppError> {
-    let storage_layer = &ctx.storage_layer;
-    let airtable = &ctx.airtable;
-    let nats = &ctx.nats;
+    let storage_layer = &services.storage_layer;
+    let airtable = &services.airtable;
 
     let current_time = Utc::now();
     let time_only = current_time.format("%H:%M:%S").to_string();
@@ -84,21 +84,16 @@ pub async fn import_airtable_base(
 
     log::info!("Started import job {job_id}");
 
-    let mut subscriber = nats.subscribe(format!("pantheon.import.cancel.{}", job_id)).await?;
-    subscriber.unsubscribe_after(1).await?;
+    let params = ImportParams {
+        name: payload.name,
+        description: payload.description,
+        job_id,
+        base_id: base_id.clone(),
+        tables: schema.tables,
+    };
 
     task::spawn(async move {
-        let _ = import_airtable_base_task(
-            ctx,
-            ImportAirtableBaseTaskParams {
-                name: payload.name,
-                description: payload.description,
-                base_id,
-                job_id,
-                subscriber,
-            },
-        )
-        .await;
+        let _ = import_task(&services, &params).await;
     });
 
     Ok(api_response::success(
