+++ src/services/storage/queries/volunteers/fetch_exported_volunteer_details_by_project_cycle.sql
@@ -1,0 +1,14 @@
+select
+  id,
+  created_at,
+  updated_at,
+  volunteer_id,
+  workspace_email,
+  org_unit,
+  job_id,
+  project_cycle_id,
+  status
+from
+  exported_volunteer_details
+where
+  project_cycle_id = $1
